feature/refactor: multiple config support
Use -c filename to source a file (must exist in the same directory as
the script), containing overrides for config variables.

All variables are still originally exported in the main script,
something that is also useful to set defaults there.

Another addition is a new command line parser, that was also needed for
the changes in the following commits.