feature/refactor: merge tests, new commands
Issues solved

- #3: Merge back test-retention and main scripts
- #5: add list and remove operations

New features

- Add rotation handlers
The "clean" function (now renamed to "check_retention") is simplified to
function as a filter, receiving on standard input a list of snapshot
information. It then checks each entry and outputs it back with a mark that
signals its status as "rotate this snapshot" or "don't rotate". By
default the function only outputs those snapshots that are marked, but
the --all option can override this behaviour.

Rotation handlers are functions that receive input as it's sent from
check_retention, and act accordingly. They can just report information,
delete snapshots that are marked for rotation, compress them, btrfs-send
them to another store, or anything one can imagine.

Handlers implemented at this point:

rotate_remove()
Remove the snapshot (what was before done by clean())

rotate_test()
Only report back whether the snapshot was marked for rotation or not

The handlers rotate_archive(), rotate_move() and rotate_btrfs_send() are
left as stubs for later implementation.

- New find() function
This one filters the list of snapshots through a regular expression, and
its output is used for many things:
  * the "find" user command
  * the "remove" user command, along with rotate_remove and its --force
    option
  * the "rotate" user command

A tokenize_find() helper is added so find()'s output (which by default
is the snapshots's full path) is formatted as tokens for the rotation
handlers and check_retention. This helper can also use a "--mark n"
option, to add a rotation mark to the stream.

- New test() function
This replaces the "random snapshot name generator" that was the central
part of the test_retention script, replacing it. It feeds its output to
check_retention and then to the rotate_test handler.

- New msg() function
All informational "echos" are done through this function, that accepts
strings both as parameters and on standard input. It then outputs to
standard error, helping in using the script output in external
pipelines.

Centralizing script output also adds possibilities in the future in
formatting, verbosity controls, and so on.