improve: snap destination/base handling
The code that determined the base directory for the current snapshot was
a bit confuse, mixing the creation of an initial snapshot with the
initialization of the $base local variable.

What's more, the code didn't allow for the user to be able to create or
have the destination already created before starting to create
snapshots.

Changes:
- initialize $base from $destination right at declaration. In case there
aren't existing subvolumes to take a snapshot from, the base will be
directly the same as the destination, otherwise it will point to the
last subvolume.

- simplify conditional

- add automatic creation of the destination directory if it doesn't exist