fix: allow $SUDO_USER access to the snap leaves
There are situations where the user is running the script as root via
sudo, as opposed to only giving sudo access to the script when it
requests it (as of now only for some btrfs subcommands).

This may be because of a need to backup directories the user doesn't
own, or for any other cause.

In such case, the *leaves* with the timestamp that the script creates
under the destination to hold the new snapshot, should be owned by the
real user. If the user doesn't have access to some rsynced object, the
permissions on that object should handle that access originally.

This also impacts favorably other operations that are based on _find()
(which at this point are almost all).