fix: exclude $destination from snapshots
The destination directory being included in the snapshots is a
compromise between a policy of "leaving all choices to the user", and a
default behaviour that was:

- patchy in code (related to the previous base/destination patch too)
- breaking least surprise by leaving at times partial base paths from
  snapshots - this didn't happen always though, but I managed to
  reproduce it with a relative $source

the policy is from now:

- OMIT by default $destination from the rsync at the sending side

- overriding of this is possible simply by adding an initial filter
  rule to clear the current ruleset (which at the point of parsing the
  user-specified rules only consists of the $destination exclusion). As
  per the rsync manual this consists of a single ! character.

Note that the use cases that require including $destination in the
snapshots are likely those that use a $destination that is used for
other content besides btrfs-backup snapshots. Now the behaviour by
default is to not favor this setup (that might lead also to errors when
detecting last snapshot!), but to allow it to be easily requested
anyway.