fix: subvol listing bug in cmd_snap()
The automatic $destination creation introduced in recent patches
introduced a regression that effectively made any subsequent snapshot
subvolume to be derived always from the root $destination subvol instead
of the last snapshot made.

This was caused by cmd_snap using btrfs subvolume list -s (filtering
only snapshots) to detect which snapshots if any existed under the
destination store.

The first time that listing produces the expected result, and an initial
snapshot is made on what is really an initial subvolume (there's nowhere
to point a snapshot to, yet). However, the next time we do want to
create a snapshot against the initial subvolume, and -s prevents that
initial subvolume to appear.

Fix this by using only the -o filter (any subvolume OR snapshot under
$destination)