fix: preserve heredocs in msg()
When using the stdin mode in msg(), passing the stream to read/echo -e
strips away leading spaces (for example) on otherwise well-formed here
documents.

As there's still not an instance of stdin mode that really needs -e,
revert to a simple cat, when the needs comes up this can be revisited.