fix: race condition in cmd_snap()
When calling cmd_snap() quickly enough (< 1s), two identical serial
numbers can generated for different snapshots, which leads to store
the second snapshot under the first, and base the second on the first.

This will in turn affect most of the rest of the operations.

Though this has a low chance of happening even in testing scenarios,
account for it refusing to create a snapshot with a serial that already
exists in the store.