feature: bitmapped tags
Add the basic constructs to have tags that encode information in
non-zero values, namely an OR bitmap of the /type/ of the tag.

A basic list of types is: leaf/day/week/month/year (for snapshots
tagged by an expiring policy rule), and manual/test types.

Add tag_info() to output back the types encoded in a tag

This in principle will be used mostly for extra information, but in
future releases enables to have per-tag-type rotation behaviors, etc.