fix: force naked file names to omit path
The top conditional in get_config was a bit naive, only checking that
the config filename that was passed didn't start with . or / as a sign
that a file with no path information was specified.

The problem is, that will happily accept file specifications that
include dots or slashes *at any other position*, which might not be
intended. This should be specially taken care of, becase config files
are sourced shell code that will potentially run as root.

A simple example:

1) identify or create a subdirectory under one of the entries in
   $config_locations (typically includes "." - the current directory)

2) pass a file with the name "subdirectory/../../some_other_path"

The file included at that point wouldn't actually be under the
intended config location, but at some other arbitrary place.

At the same time, as rare an occurrence as it might be, we need to take
into account that dots are legal characters in file names, and slashes
too (they can be escaped). So string-based checks are not much of a help
here.

The safe solution is to call basename on the file as it was passed. This
will also invalidate the otherwise legal "subdir_under_cwd/somepath"
specification, but that is easily solved by prepending ./ to the path.

The search code is slightly optimized as well to exit directly when a
file is found.