fix: reset init arrays
The multiple calls for "init basics" are intended to override the
config_file setting of (so far) config_locations, config_suffixes
and rsyncflag_progress, hinting that these arrays should only be
modified at init(). This is due to the semantics of
config_locations and config_suffixes.

However, init() wasn't resetting the arrays, merely adding the entries
with "+=". This caused repeated entries.

Thanks to the "dumpcfg" command for helping me spotting this :)